work on bugs
1. Исправил код с JDBC ( Connection,Statment,ResultSet) закрываются в коде
2. Сделал 2 класса. Считывание проперти из проперти файла. Конфигурация базы данных. Раньше все было в 1 классе
3. Переименовал класс с LiquidBaseInitialize -> LiquidBaseInitializer
4. Вынес SQL запросы в константы класса
5. Сделал Builder внутренними классами + использовал аннотацию от ломбок builder
6. Вынес PostgresContainer в отдельный класс ( для тестов )
7. Добавил remarks в liquidBase xmlфайлы

Осталось:
1. Собрать приложение полностью в Docker (Postgres,LiquidBase,Application)
2. Создать  приватную схему, для служебных таблиц
3. Поправить файл с логами, чтобы не падал в git

+++ homework/src/main/java/carshop_service/service/CarServiceImpl.java
@@ -2,36 +2,51 @@ package carshop_service.service;
 
 import carshop_service.dao.CarDao;
 import carshop_service.entity.Car;
+import carshop_service.exception.DataBaseEmptyException;
+import carshop_service.exception.DuplicateEntityException;
+import carshop_service.exception.NoSuchEntityException;
 import lombok.AllArgsConstructor;
 import java.util.List;
 
+/**
+ * Промежуточное звено между приложением и dao
+ * Обрабатывает ошибки, которые могут возникнуть при работе с car dao
+ */
 @AllArgsConstructor
 public class CarServiceImpl implements CarService {
 
-    private CarDao carDao;
+    private final CarDao carDao;
 
     @Override
-    public List<Car> getAllCars() {
-        return carDao.getAllCars();
+    public List<Car> getAllCars() throws DataBaseEmptyException {
+        List<Car> cars = carDao.getAllCars();
+        if(cars.isEmpty()) throw new DataBaseEmptyException("Car");
+        return cars;
     }
 
     @Override
-    public void addCar(Car car) {
+    public void addCar(Car car) throws DuplicateEntityException {
+        if(carDao.getCar(car.getId()) != null) throw new DuplicateEntityException("Car");
         this.carDao.addCar(car);
     }
 
     @Override
-    public Car getCar(int id) {
-        return this.carDao.getCar(id);
+    public Car getCar(int id) throws NoSuchEntityException {
+        Car car = carDao.getCar(id);
+        if(car == null) throw new NoSuchEntityException("Car");
+        return car;
     }
 
     @Override
-    public void deleteCar(int id) {
+    public void deleteCar(int id) throws NoSuchEntityException {
+        Car car = carDao.getCar(id);
+        if(car == null) throw new NoSuchEntityException("Car");
         this.carDao.deleteCar(id);
     }
 
     @Override
-    public void updateCar(Car car) {
+    public void updateCar(Car car) throws NoSuchEntityException {
+        if(carDao.getCar(car.getId()) == null) throw new NoSuchEntityException("Car");
         this.carDao.updateCar(car);
     }
